Add pg_text_to_string utility to convert PostgreSQL text* to std::string and free memory
