+++ src/utils.cpp
@@ -1,9 +1,3 @@
-extern "C" {
-#include <postgres.h>
-
-#include <utils/builtins.h>
-}
-
 #include "./include/utils.hpp"
 
 #include <algorithm>
@@ -19,20 +13,6 @@ extern "C" {
 #include "./include/logger.hpp"
 
 namespace pg_ai::utils {
-
-std::string pg_text_to_string(const text* t) {
-  if (!t)
-    return "";
-  char* cstr = text_to_cstring(t);
-  try {
-    std::string result(cstr);
-    pfree(cstr);
-    return result;
-  } catch (...) {
-    pfree(cstr);
-    throw;
-  }
-}
 
 std::pair<bool, std::string> read_file(const std::string& filepath) {
   std::ifstream file(filepath, std::ios::binary | std::ios::ate);
